Add files via upload
Found and fixed a couple typos in taxp_conversions. Changed the pums_scatter to only run conversions on TAXP. Changed the ylim and the tick style to better match the example image.

Changed the color or axs[1,1] to 'seismic' with an alpha of 0.2. I am fairly certain that color-wise, this is correct to the example image, but the different segments in the color bar are much more all over the place.

There is still a blip around 4000 on the y-axis that doesn't appear in the example image. I couldn't find the reasoning for that.